First year programming project
Worked in a team of four developing a java application for the rugby
world cup. This upload only contains a quiz section as the others are
missing. This project was a good introduction into object oriented
programming and improved my understanding and capability when working
with java and java related platforms (i.e. netbeans).